requirements.txt
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0' 
pytest
pylint
jupyter
pytest-cov
pandas
nbval
click
flask
requests
google-cloud-bigquery
google-cloud-automl

